+++ hata/discord/message/message_interaction.py
@@ -21,10 +21,12 @@ class MessageInteraction(DiscordEntity):
         The invoked interaction's name.
     type : ``InteractionType``
         The interaction's type.
+    sub_command_name_stack : `None`, `tuple` of `str`
+        The sub-command-group and sub-command names.
     user : ``ClientUserBase``
         Who invoked the interaction.
     """
-    __slots__ = ('name', 'type', 'user')
+    __slots__ = ('name', 'type', 'sub_command_name_stack', 'user')
     
     def __new__(cls, data, guild_id):
         """
@@ -37,9 +39,31 @@ class MessageInteraction(DiscordEntity):
         guild_id : `int`
             The respective message's guild's identifier.
         """
+        full_name = data.get('name', None)
+        if full_name is None:
+            name = ''
+            sub_command_name_stack = None
+        
+        else:
+            name_split = full_name.split(' ')
+            
+            name_split_length = len(name_split)
+            if name_split_length == 1:
+                name = name_split[0]
+                sub_command_name_stack = None
+            
+            elif name_split_length > 1:
+                name = name_split[0]
+                sub_command_name_stack = tuple(name_split[1:])
+            
+            else:
+                name = ''
+                sub_command_name_stack = None
+        
         self = object.__new__(cls)
         self.id = int(data['id'])
-        self.name = data['name']
+        self.name = name
+        self.sub_command_name_stack = sub_command_name_stack
         self.type = InteractionType.get(data['type'])
         self.user = User.from_data(data['user'], data.get('member', None), guild_id)
         
@@ -58,9 +82,31 @@ class MessageInteraction(DiscordEntity):
         
         repr_parts.append(', name=')
         repr_parts.append(repr(self.name))
-        repr_parts.append('>')
         
+        sub_command_name_stack = self.sub_command_name_stack
+        if (sub_command_name_stack is not None):
+            repr_parts.append(', sub_command_name_stack=')
+            repr_parts.append(repr(sub_command_name_stack))
+        
+        repr_parts.append('>')
         return ''.join(repr_parts)
+    
+    
+    @property
+    def joined_name(self):
+        """
+        Returns the joined name of the message interaction.
+        
+        Returns
+        -------
+        joined_name : `str`
+        """
+        name = self.name
+        sub_command_name_stack = self.sub_command_name_stack
+        if (sub_command_name_stack is None):
+            return name
+        
+        return ' '.join([name, *sub_command_name_stack])
     
     
     def to_data(self):
@@ -73,7 +119,7 @@ class MessageInteraction(DiscordEntity):
         """
         return {
             'id': str(self.id),
-            'name': str(self.name),
+            'name': self.joined_name,
             'type': self.type.value,
             'user': self.user.to_data()
         }